refactor(colors.rs): just break the file into multiple module
