+++ src/lib.rs
@@ -1,3 +1,7 @@
 mod ansi_escape_code;
-
+mod styled_text;
+mod styles;
 // =======================================================================
+
+pub use styled_text::styled;
+// use styled_text::StyledText;
