--- src/traits.rs
@@ -1,5 +1,0 @@
-use crate::colors::PaintType;
-
-pub trait Stylify {
-    fn make_styles(&self, paint_type: Option<&PaintType>) -> String;
-}
